% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadbandAD.R
\name{deadbandAD}
\alias{deadbandAD}
\title{deadbandAD Function}
\usage{
deadbandAD(x, EUmax, EUmin, d, offset)
}
\arguments{
\item{x}{The vector of the samples before the deadband algorithm}

\item{EUmax}{The Engineering Unit higher bound}

\item{EUmin}{The Engineering Unit lower bound}

\item{d}{Deadband percent parameter in range 0..1}

\item{offset}{How many sample do you want skip at begin? Defaults is n=20}
}
\value{
A list containing the L2 distance and the Number of filtered samples
}
\description{
This function allows you to compute the Absolute Deadband(AD) algorithm
}
\examples{
deadbandAD(rnorm(40, mean = 0, sd = 1),+0.5,-0.5,0.01,20)
}
\keyword{AD}

