% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadbandBD.R
\name{deadbandBD}
\alias{deadbandBD}
\title{deadbandBD Function}
\usage{
deadbandBD(x, d, offset, k)
}
\arguments{
\item{x}{The vector of the samples before the deadband algorithm}

\item{d}{Deadband percent parameter in range 0..1}

\item{offset}{How many sample do you want skip at begin? Defaults is n=20}

\item{k}{multiplier used in Bollinger theory}
}
\value{
A list containing the L2 distance and the Number of filtered samples
}
\description{
This function allows you to compute the Bollinger Deadband(BD) algorithm
}
\examples{
deadbandBD(rnorm(40, mean = 0, sd = 1),0.01,20,2)
}
\keyword{BD}

