% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadbandVD.R
\name{deadbandVD}
\alias{deadbandVD}
\title{deadbandVD Function}
\usage{
deadbandVD(x, d, offset, k)
}
\arguments{
\item{x}{The vector of the samples before the deadband algorithm}

\item{d}{Deadband percent parameter in range 0..1}

\item{offset}{How many sample do you want skip at begin? Defaults is n=20}

\item{k}{multiplier used in Bollinger theory}
}
\value{
A list containing the L2 distance and the Number of filtered samples
}
\description{
This function allows you to compute the Volatility Deadband(VD) algorithm
}
\examples{
deadbandVD(rnorm(40, mean = 0, sd = 1),0.01,20,2)
}
\keyword{VD}

