% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic.sub40}
\alias{synthetic.sub40}
\title{Samples subset of 10 pesudo periodic signals}
\format{A data table with a column for each signal:}
\usage{
synthetic.sub40
}
\description{
Sampling rate: 240ms for synthetic.sub40;
}
\details{
The original dataset containing the 10 pseudo periodoc signal
are available for download at:
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/synthetic-mld/synthetic.data.gz}
More Info at:
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/synthetic-mld/synthetic.data.html}

Dataset freely available for research use.
}
\keyword{datasets}

