% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic.sub42}
\alias{synthetic.sub42}
\title{Samples subset of 10 pesudo periodic signals}
\format{A data table with a column for each signal:}
\usage{
synthetic.sub42
}
\description{
Sampling rate: 252ms for synthetic.sub42;
The original dataset containing the 10 pseudo periodoc signal
are available for download at:
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/synthetic-mld/synthetic.data.gz}
More Info at:
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/synthetic-mld/synthetic.data.html}
}
\details{
Dataset freely available for research use.
}
\keyword{datasets}

